+++ Client/src/Components/Search/SearchDatasetsForm.tsx
@@ -5,9 +5,9 @@ import { MuiSelectFormik, MuiTextFieldFormik } from '../Forms/FormikFields'
 import React, { useEffect, useState } from 'react'
 
 import { MaterialSelectChipArray } from '../DatasetUpload/MetaSection/MaterialSelectChipArray'
-import { listCategories } from '../../Remote/Endpoints/CategoryEndpoint2'
-import { listMaterials } from '../../Remote/Endpoints/MaterialEndpoint2'
-import { listSubcategories } from '../../Remote/Endpoints/SubcategoryEndpoint2'
+import { listCategories } from '../../Remote/Endpoints/CategoryEndpoint'
+import { listMaterials } from '../../Remote/Endpoints/MaterialEndpoint'
+import { listSubcategories } from '../../Remote/Endpoints/SubcategoryEndpoint'
 
 interface IProps {
   handleSubmit(formValues: ISearchDatasetsFormModel): void
