+++ Client/src/Remote/Endpoints/CategoryEndpoint.tsx
